Updating old tests to use lsp, testing
